5.23(1) 重新移植H7FreeRTOS JY901S串口接收

+++ Hardware/encoder.h
@@ -1,7 +1,7 @@
 #ifndef __ENCODER_H
 #define __ENCODER_H
 
-void Get_Encoder(float* countennum1, float* countennum2);
+void Get_Encoder(short* countennum1, short* countennum2);
 
 #endif
